Changes for days!
Durability support added for block and item based restrictions.
Support for adding all blocks and items with the same damage value and
all entities.
noSpawn requires the usage of the Enum names provided in Bukkit's
EntityType.class.
Config now stored as ItemStacks and EntityTypes for their respective
options.
Removed a lot of comments that are self explanatory.
Merged Commands, ErrorMessages, Messages, and WarningMessages back into
Constants.
BlockListener, EntityListener, and PlayerListener have been rewritten
and are now more efficient (but of course despite all of that, the
double post bug with the TNT Minecart check is still there).
PlayerChecks has been deleted because of the PlayerListener's rewriting.
All of the warning messages no longer contain the plugin's prefix. It is
now inserted in the warnStaff method.
isInventoryEmpty got check automatically returns false if the inventory
has anything in it.
New method for obtaining item durabilities (will unfortunately break
with every update until a new method is found).